Refactor getDetails in hierarchy and start working on documentation
Changed getDetails() in the hierarchy to remove the thing about the
separator character; Next step will be refactor the rest of the code to
work with this changes.

Started working on documentation with doxygen. The Doxyfile for
configuration is committed, but the generated documentation is not.
The documentation inside the code should also work as general
documentation for the project, so not much more must be said in the report.